api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloud-sql-python-connector==1.4.1
cloudevents==1.10.0
cryptography==41.0.4
deprecation==2.1.0
Flask==2.3.3
frozenlist==1.4.0
functions-framework==3.4.0
google-auth==2.23.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
pg8000==1.30.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.5
watchdog==3.0.0
Werkzeug==3.0.0
yarl==1.9.2
